tests/requirements.txt
pytest
requests
paho-mqtt
psycopg2-binary
sseclient-py
pandas
